refactor(infrastructure): migrate CacheLayer to koduck-infrastructure
- Moved CacheLayer interface from koduck-core to koduck-infrastructure
- Moved RedisCacheLayer implementation from koduck-core to koduck-infrastructure
- Updated import in UserCacheServiceImpl to use new package
- Created ADR-0133 documenting the infrastructure repository implementation

This change consolidates caching infrastructure in the infrastructure module,
following the architecture decision in ADR-0131.

Closes #589
